+++ src/worker_apis/frontend_api.rs
@@ -381,7 +381,7 @@ pub async fn create_browser_login_flow(
         uri_str.add_query(&mut is_first_query, "identity_schema=", &str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -393,7 +393,9 @@ pub async fn create_browser_login_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -460,7 +462,7 @@ pub async fn create_browser_logout_flow(
         uri_str.push_str(&str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -472,7 +474,9 @@ pub async fn create_browser_logout_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -538,7 +542,7 @@ pub async fn create_browser_recovery_flow(
         uri_str.push_str(&str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -547,7 +551,9 @@ pub async fn create_browser_recovery_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -634,7 +640,7 @@ pub async fn create_browser_registration_flow(
         uri_str.add_query(&mut is_first_query, "identity_schema=", &str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -643,7 +649,9 @@ pub async fn create_browser_registration_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -711,7 +719,7 @@ pub async fn create_browser_settings_flow(
         uri_str.push_str(&str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -723,7 +731,9 @@ pub async fn create_browser_settings_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -789,7 +799,7 @@ pub async fn create_browser_verification_flow(
         uri_str.push_str(&str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -798,7 +808,9 @@ pub async fn create_browser_verification_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -857,7 +869,7 @@ pub async fn create_fedcm_flow(
     uri_str.push_str(&configuration.base_path);
     uri_str.push_str("/self-service/fed-cm/parameters");
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -866,7 +878,9 @@ pub async fn create_fedcm_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -926,7 +940,7 @@ pub async fn disable_my_other_sessions(
     uri_str.push_str(&configuration.base_path);
     uri_str.push_str("/sessions");
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -941,7 +955,9 @@ pub async fn disable_my_other_sessions(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1004,7 +1020,7 @@ pub async fn disable_my_session(
     uri_str.push_str("/sessions/");
     uri_str.push_str(&id);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1019,7 +1035,9 @@ pub async fn disable_my_session(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1071,7 +1089,7 @@ pub async fn exchange_session_token(
     uri_str.push_str("&return_to_code=");
     uri_str.push_str(&return_to_code);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1080,7 +1098,9 @@ pub async fn exchange_session_token(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1143,7 +1163,7 @@ pub async fn get_flow_error(
     uri_str.push_str("?id=");
     uri_str.push_str(&id);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1152,7 +1172,9 @@ pub async fn get_flow_error(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1216,7 +1238,7 @@ pub async fn get_login_flow(
     uri_str.push_str("?id=");
     uri_str.push_str(&id);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1228,7 +1250,9 @@ pub async fn get_login_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1294,7 +1318,7 @@ pub async fn get_recovery_flow(
     uri_str.push_str("?id=");
     uri_str.push_str(&id);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1306,7 +1330,9 @@ pub async fn get_recovery_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1372,7 +1398,7 @@ pub async fn get_registration_flow(
     uri_str.push_str("?id=");
     uri_str.push_str(&id);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1384,7 +1410,9 @@ pub async fn get_registration_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1451,7 +1479,7 @@ pub async fn get_settings_flow(
     uri_str.push_str("?id=");
     uri_str.push_str(&id);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1466,7 +1494,9 @@ pub async fn get_settings_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1532,7 +1562,7 @@ pub async fn get_verification_flow(
     uri_str.push_str("?id=");
     uri_str.push_str(&id);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1544,7 +1574,9 @@ pub async fn get_verification_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1602,7 +1634,7 @@ pub async fn get_web_authn_java_script(
     uri_str.push_str(&configuration.base_path);
     uri_str.push_str("/.well-known/ory/webauthn.js");
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1611,7 +1643,9 @@ pub async fn get_web_authn_java_script(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1696,7 +1730,7 @@ pub async fn list_my_sessions(
         uri_str.add_query(&mut is_first_query, "page_token=", &str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1711,7 +1745,9 @@ pub async fn list_my_sessions(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1778,7 +1814,7 @@ pub async fn to_session(
         uri_str.push_str(&str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1793,7 +1829,9 @@ pub async fn to_session(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -1851,7 +1889,7 @@ pub async fn update_fedcm_flow(
     uri_str.push_str(&configuration.base_path);
     uri_str.push_str("/self-service/fed-cm/token");
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1860,7 +1898,9 @@ pub async fn update_fedcm_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
     headers.append("Content-Type", "application/json")?;
 
@@ -1928,7 +1968,7 @@ pub async fn update_login_flow(
     uri_str.push_str("?flow=");
     uri_str.push_str(&flow);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -1943,7 +1983,9 @@ pub async fn update_login_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
     headers.append("Content-Type", "application/json")?;
 
@@ -2019,7 +2061,7 @@ pub async fn update_logout_flow(
         uri_str.add_query(&mut is_first_query, "token=", &str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -2031,7 +2073,9 @@ pub async fn update_logout_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
 
     let mut req_builder = RequestInit::new();
@@ -2087,7 +2131,7 @@ pub async fn update_recovery_flow(
         uri_str.push_str(&str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -2099,7 +2143,9 @@ pub async fn update_recovery_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
     headers.append("Content-Type", "application/json")?;
 
@@ -2168,7 +2214,7 @@ pub async fn update_registration_flow(
     uri_str.push_str("?flow=");
     uri_str.push_str(&flow);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -2180,7 +2226,9 @@ pub async fn update_registration_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
     headers.append("Content-Type", "application/json")?;
 
@@ -2250,7 +2298,7 @@ pub async fn update_settings_flow(
     uri_str.push_str("?flow=");
     uri_str.push_str(&flow);
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -2265,7 +2313,9 @@ pub async fn update_settings_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
     headers.append("Content-Type", "application/json")?;
 
@@ -2341,7 +2391,7 @@ pub async fn update_verification_flow(
         uri_str.push_str(&str);
     }
 
-    let mut headers = Headers::new();
+    let headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -2353,7 +2403,9 @@ pub async fn update_verification_flow(
     headers.append("Accept", "application/json")?;
 
     if let Some(ref cors) = configuration.cors {
-        cors.apply_headers(&mut headers)?;
+        for (k, v) in cors.entries() {
+            headers.append(&k, &v)?;
+        }
     }
     headers.append("Content-Type", "application/json")?;